pyproject.toml
[tool.poetry]
name = "event-scrapping"
version = "0.1.0"
description = ""
authors = ["paras"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.7"
flake8 = "^7.1.0"
isort = "^5.13.2"
black = "^24.4.2"
djangorestframework = "^3.15.2"
requests = "^2.32.3"
pytz = "^2024.1"
xmltodict = "^0.13.0"
django-apscheduler = "^0.6.2"
apscheduler = "^3.10.4"
django-environ = "^0.11.2"
requests-mock = "^1.12.1"
dj-database-url = "^2.2.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
pytest = "^8.2.2"
pytest-django = "^4.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
